requirements.txt
Django
Django-REST-framework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
python-telegram-bot
Pillow
apscheduler
whitenoise
gunicorn==20.1.0
django_apscheduler
requests
python-dotenv
supabase
pandas
openpyxl
